Move compliance tooling into internal package
