+++ rclip/_compliance/__init__.py
@@ -1,0 +1,1 @@
+"""Internal build-time licence and distribution compliance support."""
