--- rclip/compliance/__init__.py
@@ -1,1 +1,0 @@
-"""Build-time licence and distribution compliance support."""
