+++ tests/unit/test_compliance.py
@@ -7,18 +7,18 @@ import tomllib
 
 import pytest
 
-from rclip.compliance.common import ComplianceError
-from rclip.compliance.legal import collect_legal_materials
-from rclip.compliance.licenses import _declared_license_expression
-from rclip.compliance.licenses import _validate_python_packages
-from rclip.compliance.native import _binary_contains
-from rclip.compliance.native import _native_component_versions
-from rclip.compliance.policy import _locked_runtime_versions
-from rclip.compliance.policy import load_policy
-from rclip.compliance.sbom import augment_cyclonedx
-from rclip.compliance.source import _deterministic_tar
-from rclip.compliance.source import build_corresponding_source
-from rclip.compliance.verify import verify_bundle
+from rclip._compliance.common import ComplianceError
+from rclip._compliance.legal import collect_legal_materials
+from rclip._compliance.licenses import _declared_license_expression
+from rclip._compliance.licenses import _validate_python_packages
+from rclip._compliance.native import _binary_contains
+from rclip._compliance.native import _native_component_versions
+from rclip._compliance.policy import _locked_runtime_versions
+from rclip._compliance.policy import load_policy
+from rclip._compliance.sbom import augment_cyclonedx
+from rclip._compliance.source import _deterministic_tar
+from rclip._compliance.source import build_corresponding_source
+from rclip._compliance.verify import verify_bundle
 
 
 REPO_ROOT = Path(__file__).resolve().parents[2]
@@ -138,7 +138,7 @@ def test_does_not_collect_python_runtime_license_twice(tmp_path: Path, monkeypat
   python_license.parent.mkdir(parents=True)
   python_license.write_text("CPython licence\n", encoding="utf-8")
   monkeypatch.setattr(
-    "rclip.compliance.legal._python_runtime_license",
+    "rclip._compliance.legal._python_runtime_license",
     lambda _root: ("3.11.0", python_license),
   )
   output = root / "usr/share/doc/rclip"
@@ -158,7 +158,7 @@ def test_omits_external_python_runtime_by_default(tmp_path: Path, monkeypatch: p
   root = tmp_path / "runtime"
   write_distribution(root, "rclip", "3.3.0")
   monkeypatch.setattr(
-    "rclip.compliance.legal._python_runtime_license",
+    "rclip._compliance.legal._python_runtime_license",
     lambda _root: pytest.fail("external Python runtime should not be inspected"),
   )
 
@@ -362,7 +362,7 @@ def test_rejects_enabled_or_unreported_rawpy_gpl_features(
       self.flags = feature_flags
 
   rawpy = Rawpy(flags)
-  monkeypatch.setattr("rclip.compliance.native.importlib.import_module", lambda name: rawpy)
+  monkeypatch.setattr("rclip._compliance.native.importlib.import_module", lambda name: rawpy)
 
   with pytest.raises(ComplianceError, match=error):
     _native_component_versions({"rawpy"})
